fix(core): 修复二进制类型
- 字节频率在文件超过约 4 GB 后会造成溢出，因此需要改为 u64，最大支持 16 EB（几乎不可能）。
- 频率表项数量改为 u16 以节省二进制空间，不用 u8 是因为 256 时正好会溢出为 0，与空文件冲突。
- 修改针对频次表的测试代码。

+++ crates/huffman_core/src/format/v1.rs
@@ -108,7 +108,7 @@ pub struct FrequencyEntry {
     pub symbol: u8,
 
     /// 字符出现频率
-    pub frequency: u32,
+    pub frequency: u64,
 }
 
 /// 频次表区块结构体
@@ -116,7 +116,7 @@ pub struct FrequencyEntry {
 #[brw(big)]
 pub struct FrequencyTable {
     /// 频率表项数量
-    pub count: u32,
+    pub count: u16,
 
     /// 频率表项数组
     #[br(count = count)]
